# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"responses":{"200":{"description":""}},"summary":"User Register","tags":["Auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"summary":"User Login","tags":["Auth"]}},"/api/v1/auth/get-me":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/change-password":{"post":{"operationId":"AuthController_changePassword","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPasswod","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[{"name":"authorization","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/v1/users/create-admin":{"post":{"operationId":"UsersController_createAdmin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserAdminDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/api/v1/users":{"get":{"operationId":"UsersController_getUsers","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/v1/admins":{"get":{"operationId":"AdminController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/api/v1/admins/{id}":{"get":{"operationId":"AdminController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Admin"]},"patch":{"operationId":"AdminController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAdminDto"}}}},"responses":{"200":{"description":""}},"tags":["Admin"]},"delete":{"operationId":"AdminController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/api/v1/analytics/admin":{"get":{"operationId":"AnalyticsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Analytics"]}},"/api/v1/analytics/trader":{"get":{"operationId":"AnalyticsController_adminEconomic","parameters":[],"responses":{"200":{"description":""}},"tags":["Analytics"]}},"/api/v1/webhook":{"post":{"operationId":"WebhookController_handleWebhook","parameters":[{"name":"stripe-signature","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Webhook"]}},"/api/v1/message":{"post":{"operationId":"MessageController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMessageDto"}}}},"responses":{"201":{"description":""}},"tags":["Message"]},"get":{"operationId":"MessageController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Message"]}},"/api/v1/message/{id}":{"get":{"operationId":"MessageController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Message"]},"patch":{"operationId":"MessageController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMessageDto"}}}},"responses":{"200":{"description":""}},"tags":["Message"]},"delete":{"operationId":"MessageController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Message"]}},"/api/v1/payment":{"post":{"operationId":"PaymentController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePaymentDto"}}}},"responses":{"201":{"description":""}},"tags":["Payment"]},"get":{"operationId":"PaymentController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Payment"]}},"/api/v1/payment/{id}":{"get":{"operationId":"PaymentController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payment"]},"patch":{"operationId":"PaymentController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePaymentDto"}}}},"responses":{"200":{"description":""}},"tags":["Payment"]},"delete":{"operationId":"PaymentController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payment"]}},"/api/v1/reffering":{"post":{"operationId":"RefferingController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRefferingDto"}}}},"responses":{"201":{"description":""}},"tags":["Reffering"]},"get":{"operationId":"RefferingController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Reffering"]}},"/api/v1/reffering/{id}":{"get":{"operationId":"RefferingController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Reffering"]},"patch":{"operationId":"RefferingController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRefferingDto"}}}},"responses":{"200":{"description":""}},"tags":["Reffering"]},"delete":{"operationId":"RefferingController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Reffering"]}},"/api/v1/review":{"post":{"operationId":"ReviewController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"201":{"description":""}},"tags":["Review"]},"get":{"operationId":"ReviewController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Review"]}},"/api/v1/review/{id}":{"get":{"operationId":"ReviewController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Review"]},"patch":{"operationId":"ReviewController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateReviewDto"}}}},"responses":{"200":{"description":""}},"tags":["Review"]},"delete":{"operationId":"ReviewController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Review"]}},"/api/v1/subscription":{"post":{"operationId":"SubscriptionController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubscriptionDto"}}}},"responses":{"201":{"description":""}},"tags":["Subscription"]},"get":{"operationId":"SubscriptionController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Subscription"]}},"/api/v1/subscription/{id}":{"get":{"operationId":"SubscriptionController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Subscription"]},"patch":{"operationId":"SubscriptionController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSubscriptionDto"}}}},"responses":{"200":{"description":""}},"tags":["Subscription"]},"delete":{"operationId":"SubscriptionController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Subscription"]}},"/api/v1/task-management":{"post":{"operationId":"TaskManagementController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskManagementDto"}}}},"responses":{"201":{"description":""}},"tags":["TaskManagement"]},"get":{"operationId":"TaskManagementController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["TaskManagement"]}},"/api/v1/task-management/{id}":{"get":{"operationId":"TaskManagementController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["TaskManagement"]},"patch":{"operationId":"TaskManagementController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskManagementDto"}}}},"responses":{"200":{"description":""}},"tags":["TaskManagement"]},"delete":{"operationId":"TaskManagementController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["TaskManagement"]}},"/api/v1/customers":{"get":{"operationId":"CustomerController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Customer"]}},"/api/v1/customers/{id}":{"get":{"operationId":"CustomerController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Customer"]},"patch":{"operationId":"CustomerController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCustomerDto"}}}},"responses":{"200":{"description":""}},"tags":["Customer"]},"delete":{"operationId":"CustomerController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Customer"]}},"/api/v1/verification":{"post":{"operationId":"VerificationController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVerificationDto"}}}},"responses":{"201":{"description":""}},"tags":["Verification"]},"get":{"operationId":"VerificationController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Verification"]}},"/api/v1/verification/{id}":{"get":{"operationId":"VerificationController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Verification"]},"patch":{"operationId":"VerificationController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateVerificationDto"}}}},"responses":{"200":{"description":""}},"tags":["Verification"]},"delete":{"operationId":"VerificationController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Verification"]}}},"info":{"title":"kmathew95_Server","description":"The kmathew95_Server API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"LoginDto":{"type":"object","properties":{}},"CreateUserAdminDto":{"type":"object","properties":{}},"UpdateAdminDto":{"type":"object","properties":{}},"CreateMessageDto":{"type":"object","properties":{}},"UpdateMessageDto":{"type":"object","properties":{}},"CreatePaymentDto":{"type":"object","properties":{}},"UpdatePaymentDto":{"type":"object","properties":{}},"CreateRefferingDto":{"type":"object","properties":{}},"UpdateRefferingDto":{"type":"object","properties":{}},"CreateReviewDto":{"type":"object","properties":{}},"UpdateReviewDto":{"type":"object","properties":{}},"CreateSubscriptionDto":{"type":"object","properties":{}},"UpdateSubscriptionDto":{"type":"object","properties":{}},"CreateTaskManagementDto":{"type":"object","properties":{}},"UpdateTaskManagementDto":{"type":"object","properties":{}},"UpdateCustomerDto":{"type":"object","properties":{}},"CreateVerificationDto":{"type":"object","properties":{}},"UpdateVerificationDto":{"type":"object","properties":{}}}}}